refactor: reorganize file structure for admin components
