+++ package.json
@@ -33,6 +33,7 @@
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "globals": "^15.15.0",
     "postcss": "^8.5.3",
     "prettier": "^3.5.3",
