+++ src/components/admin/applicant/AdminConfirmModal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../common/Button";
+import Button from "../../../components/common/Button";
 
 interface ConfirmModalProps {
   isOpen: boolean;
